categories, regions 를 selectList 로 리팩토링
